Use a simple pseudo-random generator
This is enough for our needs and it ensures that the pseudo-randomness
always provides the same perturbations of the equidistant points.
This is interesting, say, when generating graphs for a paper because
the generated files are not constantly changing in spite of not having
modified the function.

This also enables to drop the dependency on the crate `rand`.