Attempt to represent `x` and `y` as vectors (much slower)
This is an experiment to represent the coordinates `x` and `y` as
vectors which will enable projecting on these coordinates simply
return references to the underlying data.  This version is much
slower.  Even using an "indirection vector" between the priority queue
and the indices of the path (vector that one updates when inserting a
point — which shift indices), the running time is 6–7 times slower
than the original version.  The problem is that the `insert` function
is much more costly (and it is a frequent operation).